sales and corresonding alerts dataset created
