+++ alerts.csv
@@ -1,0 +1,43 @@
+product_id,payment,time_stamp
+0,5000,7:28
+1,170000,6:59
+2,21000.0,8:12
+2,7000.0,8:35
+3,112500.0,8:49
+3,37500.0,9:21
+4,210000,8:50
+5,102000.0,8:57
+5,34000.0,9:25
+6,15000,9:56
+7,3750.0,9:51
+7,1250.0,10:22
+8,26250.0,9:46
+8,8750.0,10:12
+9,34000,11:28
+10,105000,11:27
+11,200000,13:42
+12,21000.0,13:44
+12,7000.0,14:28
+13,30000,14:25
+14,136000,14:39
+15,28000,14:34
+16,52500.0,14:29
+16,17500.0,15:20
+17,18750.0,14:37
+17,6250.0,15:27
+18,34000,15:31
+19,15000,15:5
+20,14000,15:42
+21,78750.0,16:5
+21,26250.0,16:22
+22,51000,16:34
+23,34000,17:32
+24,15000.0,17:7
+24,5000.0,17:47
+25,35000,17:36
+26,34000,18:16
+27,26250.0,18:14
+27,8750.0,18:48
+28,26250.0,19:36
+28,8750.0,20:14
+29,100000,20:23
