+++ sales.csv
@@ -1,0 +1,31 @@
+product_id,price,time_stamp
+0,5000,07:03
+1,170000,07:16
+2,28000,08:28
+3,150000,08:56
+4,210000,09:01
+5,136000,09:06
+6,15000,09:57
+7,5000,10:04
+8,35000,10:07
+9,34000,11:01
+10,105000,11:03
+11,200000,13:53
+12,28000,14:00
+13,30000,14:01
+14,136000,14:22
+15,28000,14:35
+16,70000,14:55
+17,25000,14:58
+18,34000,15:07
+19,15000,15:23
+20,14000,15:47
+21,105000,16:17
+22,51000,16:35
+23,34000,17:13
+24,20000,17:20
+25,35000,17:43
+26,34000,18:30
+27,35000,18:41
+28,35000,19:50
+29,100000,19:56
